+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add methods: `getAutoZoom`, `setAutoZoom`, `loadMusicXML`, `getMusicXML`
 * Add events: `scoreLoaded`, `cursorPosition`, `rangeSelection`, `pause`, `stop`, `fullscreen`, `print`
+* Add integration tests
 
 ## v0.1.0